+++ projects/ngx-pwa/local-storage/src/lib/validation/json-validation.spec.ts
@@ -1,4 +1,8 @@
 import { JSONValidator } from './json-validator';
+import {
+  JSONSchemaConst, JSONSchemaEnum, JSONSchemaBoolean, JSONSchemaNull,
+  JSONSchemaNumeric, JSONSchemaString, JSONSchemaObject, JSONSchemaArray
+} from './json-schema';
 
 describe(`JSONValidator`, () => {
 
@@ -10,13 +14,15 @@ describe(`JSONValidator`, () => {
 
   });
 
-  describe(`validation`, () => {
+  describe(`using an existing JSON Schema`, () => {
 
-    it(`should throw if nothing describes the value`, () => {
+    it(`should not throw with options unsupported by the lib but allowed in the JSON Schema standard`, () => {
 
-      expect(() => {
-        jsonValidator.validate('test', {});
-      }).toThrowError();
+      expect(() => {
+
+        jsonValidator.validate({ test: 'test' }, { properties: { test: { type: 'string' } }, additionalProperties: true });
+
+      }).not.toThrow();
 
     });
 
@@ -26,7 +32,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on a string equal to a string const`, () => {
 
-      const test = jsonValidator.validate('test', { const: 'test' });
+      const test = jsonValidator.validate('test', { const: 'test' } as JSONSchemaConst);
 
       expect(test).toBe(true);
 
@@ -34,7 +40,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on a string not equal to a string const`, () => {
 
-      const test = jsonValidator.validate('test2', { const: 'test' });
+      const test = jsonValidator.validate('test2', { const: 'test' } as JSONSchemaConst);
 
       expect(test).toBe(false);
 
@@ -42,7 +48,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on a number equal to a number const`, () => {
 
-      const test = jsonValidator.validate(1.5, { const: 1.5 });
+      const test = jsonValidator.validate(1.5, { const: 1.5 } as JSONSchemaConst);
 
       expect(test).toBe(true);
 
@@ -50,7 +56,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on a number not equal to a number const`, () => {
 
-      const test = jsonValidator.validate(2.5, { const: 1.5 });
+      const test = jsonValidator.validate(2.5, { const: 1.5 } as JSONSchemaConst);
 
       expect(test).toBe(false);
 
@@ -58,7 +64,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on an integer equal to an integer const`, () => {
 
-      const test = jsonValidator.validate(1, { const: 1 });
+      const test = jsonValidator.validate(1, { const: 1 } as JSONSchemaConst);
 
       expect(test).toBe(true);
 
@@ -66,7 +72,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on an integer not equal to an integer const`, () => {
 
-      const test = jsonValidator.validate(2, { const: 1 });
+      const test = jsonValidator.validate(2, { const: 1 } as JSONSchemaConst);
 
       expect(test).toBe(false);
 
@@ -74,7 +80,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on a boolean equal to a boolean const`, () => {
 
-      const test = jsonValidator.validate(true, { const: true });
+      const test = jsonValidator.validate(true, { const: true } as JSONSchemaConst);
 
       expect(test).toBe(true);
 
@@ -82,7 +88,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on a boolean not equal to a boolean const`, () => {
 
-      const test = jsonValidator.validate(false, { const: true });
+      const test = jsonValidator.validate(false, { const: true } as JSONSchemaConst);
 
       expect(test).toBe(false);
 
@@ -90,7 +96,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on null equal to a null const`, () => {
 
-      const test = jsonValidator.validate(null, { const: null });
+      const test = jsonValidator.validate(null, { const: null } as JSONSchemaConst);
 
       expect(test).toBe(true);
 
@@ -98,7 +104,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on a value not equal to a null const`, () => {
 
-      const test = jsonValidator.validate('test', { const: null });
+      const test = jsonValidator.validate('test', { const: null } as JSONSchemaConst);
 
       expect(test).toBe(false);
 
@@ -106,7 +112,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on an empty string with a false const`, () => {
 
-      const test = jsonValidator.validate('', { const: false });
+      const test = jsonValidator.validate('', { const: false } as JSONSchemaConst);
 
       expect(test).toBe(false);
 
@@ -114,7 +120,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on 0 with a false const`, () => {
 
-      const test = jsonValidator.validate(0, { const: false });
+      const test = jsonValidator.validate(0, { const: false } as JSONSchemaConst);
 
       expect(test).toBe(false);
 
@@ -122,7 +128,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on an empty string with a null const`, () => {
 
-      const test = jsonValidator.validate('', { const: null });
+      const test = jsonValidator.validate('', { const: null } as JSONSchemaConst);
 
       expect(test).toBe(false);
 
@@ -130,7 +136,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on 0 with a null const`, () => {
 
-      const test = jsonValidator.validate(0, { const: null });
+      const test = jsonValidator.validate(0, { const: null } as JSONSchemaConst);
 
       expect(test).toBe(false);
 
@@ -142,7 +148,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on a value included in an enum`, () => {
 
-      const test = jsonValidator.validate('test', { enum: ['test', 'hello'] });
+      const test = jsonValidator.validate('test', { enum: ['test', 'hello'] } as JSONSchemaEnum);
 
       expect(test).toBe(true);
 
@@ -150,7 +156,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on a value not included in an enum`, () => {
 
-      const test = jsonValidator.validate('test2', { enum: ['test', 'hello'] });
+      const test = jsonValidator.validate('test2', { enum: ['test', 'hello'] } as JSONSchemaEnum);
 
       expect(test).toBe(false);
 
@@ -158,7 +164,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on an empty string included in an enum`, () => {
 
-      const test = jsonValidator.validate('', { enum: ['', 'hello'] });
+      const test = jsonValidator.validate('', { enum: ['', 'hello'] } as JSONSchemaEnum);
 
       expect(test).toBe(true);
 
@@ -166,7 +172,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on 0 included in an enum`, () => {
 
-      const test = jsonValidator.validate(0, { enum: [0, 1] });
+      const test = jsonValidator.validate(0, { enum: [0, 1] } as JSONSchemaEnum);
 
       expect(test).toBe(true);
 
@@ -178,7 +184,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on a true value with a boolean type`, () => {
 
-      const test = jsonValidator.validate(true, { type: 'boolean' });
+      const test = jsonValidator.validate(true, { type: 'boolean' } as JSONSchemaBoolean);
 
       expect(test).toBe(true);
 
@@ -186,31 +192,15 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on a false value with a boolean type`, () => {
 
-      const test = jsonValidator.validate(false, { type: 'boolean' });
+      const test = jsonValidator.validate(false, { type: 'boolean' } as JSONSchemaBoolean);
 
       expect(test).toBe(true);
 
     });
 
-    it(`should throw on an object value with an object type without 'properties'`, () => {
-
-      expect(() => {
-        jsonValidator.validate({}, { type: 'object' });
-      }).toThrow();
-
-    });
-
-    it(`should throw on an array value with an array type without 'items'`, () => {
-
-      expect(() => {
-        jsonValidator.validate({}, { type: 'array' });
-      }).toThrow();
-
-    });
-
     it(`should return true on a null value with a null type`, () => {
 
-      const test = jsonValidator.validate(null, { type: 'null' });
+      const test = jsonValidator.validate(null, { type: 'null' } as JSONSchemaNull);
 
       expect(test).toBe(true);
 
@@ -218,7 +208,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on a primitive value with a mismatched type`, () => {
 
-      const test = jsonValidator.validate('test', { type: 'number' });
+      const test = jsonValidator.validate('test', { type: 'number' } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -230,7 +220,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on a string value with string type`, () => {
 
-      const test = jsonValidator.validate('test', { type: 'string' });
+      const test = jsonValidator.validate('test', { type: 'string' } as JSONSchemaString);
 
       expect(test).toBe(true);
 
@@ -239,7 +229,7 @@ describe(`JSONValidator`, () => {
     it(`should throw if maxLength is not an integer`, () => {
 
       expect(() => {
-        jsonValidator.validate('test', { type: 'string', maxLength: 10.5 });
+        jsonValidator.validate('test', { type: 'string', maxLength: 10.5 } as JSONSchemaString);
       }).toThrowError();
 
     });
@@ -247,14 +237,14 @@ describe(`JSONValidator`, () => {
     it(`should throw if maxLength is negative`, () => {
 
       expect(() => {
-        jsonValidator.validate('test', { type: 'string', maxLength: -1 });
+        jsonValidator.validate('test', { type: 'string', maxLength: -1 } as JSONSchemaString);
       }).toThrowError();
 
     });
 
     it(`should return true with a string respecting maxLength`, () => {
 
-      const test = jsonValidator.validate('test', { type: 'string', maxLength: 10 });
+      const test = jsonValidator.validate('test', { type: 'string', maxLength: 10 } as JSONSchemaString);
 
       expect(test).toBe(true);
 
@@ -262,7 +252,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with a string not respecting maxLength`, () => {
 
-      const test = jsonValidator.validate('test', { type: 'string', maxLength: 2 });
+      const test = jsonValidator.validate('test', { type: 'string', maxLength: 2 } as JSONSchemaString);
 
       expect(test).toBe(false);
 
@@ -271,7 +261,7 @@ describe(`JSONValidator`, () => {
     it(`should throw if minLength is not an integer`, () => {
 
       expect(() => {
-        jsonValidator.validate('test', { type: 'string', minLength: 10.5 });
+        jsonValidator.validate('test', { type: 'string', minLength: 10.5 } as JSONSchemaString);
       }).toThrowError();
 
     });
@@ -279,14 +269,14 @@ describe(`JSONValidator`, () => {
     it(`should throw if minLength is negative`, () => {
 
       expect(() => {
-        jsonValidator.validate('test', { type: 'string', minLength: -1 });
+        jsonValidator.validate('test', { type: 'string', minLength: -1 } as JSONSchemaString);
       }).toThrowError();
 
     });
 
     it(`should return true with a string respecting minLength`, () => {
 
-      const test = jsonValidator.validate('test', { type: 'string', minLength: 2 });
+      const test = jsonValidator.validate('test', { type: 'string', minLength: 2 } as JSONSchemaString);
 
       expect(test).toBe(true);
 
@@ -294,7 +284,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with a string not respecting minLength`, () => {
 
-      const test = jsonValidator.validate('t', { type: 'string', minLength: 2 });
+      const test = jsonValidator.validate('t', { type: 'string', minLength: 2 } as JSONSchemaString);
 
       expect(test).toBe(false);
 
@@ -303,14 +293,14 @@ describe(`JSONValidator`, () => {
     it(`should throw if pattern is not valid`, () => {
 
       expect(() => {
-        jsonValidator.validate('test', { type: 'string', pattern: '+++' });
+        jsonValidator.validate('test', { type: 'string', pattern: '+++' } as JSONSchemaString);
       }).toThrowError();
 
     });
 
     it(`should return true with a string respecting a pattern`, () => {
 
-      const test = jsonValidator.validate('test', { type: 'string', pattern: '^test$' });
+      const test = jsonValidator.validate('test', { type: 'string', pattern: '^test$' } as JSONSchemaString);
 
       expect(test).toBe(true);
 
@@ -318,7 +308,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with a string not respecting a pattern`, () => {
 
-      const test = jsonValidator.validate('t', { type: 'string', pattern: '^test$' });
+      const test = jsonValidator.validate('t', { type: 'string', pattern: '^test$' } as JSONSchemaString);
 
       expect(test).toBe(false);
 
@@ -330,7 +320,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on an integer value with a number type`, () => {
 
-      const test = jsonValidator.validate(10, { type: 'number' });
+      const test = jsonValidator.validate(10, { type: 'number' } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -338,7 +328,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on a decimal value with a number type`, () => {
 
-      const test = jsonValidator.validate(10.1, { type: 'number' });
+      const test = jsonValidator.validate(10.1, { type: 'number' } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -346,7 +336,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true on an integer value with an integer type`, () => {
 
-      const test = jsonValidator.validate(10, { type: 'integer' });
+      const test = jsonValidator.validate(10, { type: 'integer' } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -354,7 +344,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false on a decimal value with an integer type`, () => {
 
-      const test = jsonValidator.validate(10.1, { type: 'integer' });
+      const test = jsonValidator.validate(10.1, { type: 'integer' } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -363,14 +353,14 @@ describe(`JSONValidator`, () => {
     it(`should throw if multipleOf is not strictly greater than 0`, () => {
 
       expect(() => {
-        jsonValidator.validate(10, { type: 'number', multipleOf: 0 });
+        jsonValidator.validate(10, { type: 'number', multipleOf: 0 } as JSONSchemaNumeric);
       }).toThrowError();
 
     });
 
     it(`should return true with a number is a multiple of x`, () => {
 
-      const test = jsonValidator.validate(5.2, { type: 'number', multipleOf: 2.6 });
+      const test = jsonValidator.validate(5.2, { type: 'number', multipleOf: 2.6 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -378,7 +368,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with an integer is a multiple of x`, () => {
 
-      const test = jsonValidator.validate(10, { type: 'integer', multipleOf: 2 });
+      const test = jsonValidator.validate(10, { type: 'integer', multipleOf: 2 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -386,7 +376,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with a number not a multiple of x`, () => {
 
-      const test = jsonValidator.validate(5.3, { type: 'number', multipleOf: 2.6 });
+      const test = jsonValidator.validate(5.3, { type: 'number', multipleOf: 2.6 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -394,7 +384,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with an integer not a multiple of x`, () => {
 
-      const test = jsonValidator.validate(11, { type: 'integer', multipleOf: 2 });
+      const test = jsonValidator.validate(11, { type: 'integer', multipleOf: 2 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -402,7 +392,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with a number less than a maximum`, () => {
 
-      const test = jsonValidator.validate(5.2, { type: 'number', maximum: 5.7 });
+      const test = jsonValidator.validate(5.2, { type: 'number', maximum: 5.7 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -410,7 +400,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with a number equal to maximum`, () => {
 
-      const test = jsonValidator.validate(5.2, { type: 'number', maximum: 5.2 });
+      const test = jsonValidator.validate(5.2, { type: 'number', maximum: 5.2 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -418,7 +408,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with a number greater than a maximum`, () => {
 
-      const test = jsonValidator.validate(5.3, { type: 'number', maximum: 5.2 });
+      const test = jsonValidator.validate(5.3, { type: 'number', maximum: 5.2 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -426,7 +416,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with an integer less than a maximum`, () => {
 
-      const test = jsonValidator.validate(5, { type: 'integer', maximum: 6 });
+      const test = jsonValidator.validate(5, { type: 'integer', maximum: 6 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -434,7 +424,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with an integer equal to maximum`, () => {
 
-      const test = jsonValidator.validate(5, { type: 'integer', maximum: 5 });
+      const test = jsonValidator.validate(5, { type: 'integer', maximum: 5 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -442,7 +432,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with an integer greater than a maximum`, () => {
 
-      const test = jsonValidator.validate(6, { type: 'integer', maximum: 5 });
+      const test = jsonValidator.validate(6, { type: 'integer', maximum: 5 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -450,7 +440,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with a number less than an exclusiveMaximum`, () => {
 
-      const test = jsonValidator.validate(5.2, { type: 'number', exclusiveMaximum: 5.7 });
+      const test = jsonValidator.validate(5.2, { type: 'number', exclusiveMaximum: 5.7 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -458,7 +448,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with a number equal to exclusiveMaximum`, () => {
 
-      const test = jsonValidator.validate(5.2, { type: 'number', exclusiveMaximum: 5.2 });
+      const test = jsonValidator.validate(5.2, { type: 'number', exclusiveMaximum: 5.2 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -466,7 +456,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with a number greater than an exclusiveMaximum`, () => {
 
-      const test = jsonValidator.validate(5.3, { type: 'number', exclusiveMaximum: 5.2 });
+      const test = jsonValidator.validate(5.3, { type: 'number', exclusiveMaximum: 5.2 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -474,7 +464,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with an integer less than an exclusiveMaximum`, () => {
 
-      const test = jsonValidator.validate(5, { type: 'integer', exclusiveMaximum: 6 });
+      const test = jsonValidator.validate(5, { type: 'integer', exclusiveMaximum: 6 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -482,7 +472,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with an integer equal to exclusiveMaximum`, () => {
 
-      const test = jsonValidator.validate(5, { type: 'integer', exclusiveMaximum: 5 });
+      const test = jsonValidator.validate(5, { type: 'integer', exclusiveMaximum: 5 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -490,7 +480,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with an integer greater than an exclusiveMaximum`, () => {
 
-      const test = jsonValidator.validate(6, { type: 'integer', exclusiveMaximum: 5 });
+      const test = jsonValidator.validate(6, { type: 'integer', exclusiveMaximum: 5 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -498,7 +488,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with a number greater than a minimum`, () => {
 
-      const test = jsonValidator.validate(5.8, { type: 'number', minimum: 5.7 });
+      const test = jsonValidator.validate(5.8, { type: 'number', minimum: 5.7 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -506,7 +496,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with a number equal to minimum`, () => {
 
-      const test = jsonValidator.validate(5.2, { type: 'number', minimum: 5.2 });
+      const test = jsonValidator.validate(5.2, { type: 'number', minimum: 5.2 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -514,7 +504,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with a number less than an minimum`, () => {
 
-      const test = jsonValidator.validate(5.1, { type: 'number', minimum: 5.2 });
+      const test = jsonValidator.validate(5.1, { type: 'number', minimum: 5.2 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -522,7 +512,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with an integer greater than a minimum`, () => {
 
-      const test = jsonValidator.validate(7, { type: 'integer', minimum: 6 });
+      const test = jsonValidator.validate(7, { type: 'integer', minimum: 6 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -530,7 +520,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with an integer equal to minimum`, () => {
 
-      const test = jsonValidator.validate(5, { type: 'integer', minimum: 5 });
+      const test = jsonValidator.validate(5, { type: 'integer', minimum: 5 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -538,7 +528,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with an integer less than an minimum`, () => {
 
-      const test = jsonValidator.validate(4, { type: 'integer', minimum: 5 });
+      const test = jsonValidator.validate(4, { type: 'integer', minimum: 5 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -546,7 +536,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with a number greater than an exclusiveMinimum`, () => {
 
-      const test = jsonValidator.validate(5.8, { type: 'number', exclusiveMinimum: 5.7 });
+      const test = jsonValidator.validate(5.8, { type: 'number', exclusiveMinimum: 5.7 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -554,7 +544,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with a number equal to exclusiveMinimum`, () => {
 
-      const test = jsonValidator.validate(5.2, { type: 'number', exclusiveMinimum: 5.2 });
+      const test = jsonValidator.validate(5.2, { type: 'number', exclusiveMinimum: 5.2 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -562,7 +552,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with a number less than an exclusiveMinimum`, () => {
 
-      const test = jsonValidator.validate(5.1, { type: 'number', exclusiveMinimum: 5.2 });
+      const test = jsonValidator.validate(5.1, { type: 'number', exclusiveMinimum: 5.2 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -570,7 +560,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with an integer greater than an exclusiveMinimum`, () => {
 
-      const test = jsonValidator.validate(7, { type: 'integer', exclusiveMinimum: 6 });
+      const test = jsonValidator.validate(7, { type: 'integer', exclusiveMinimum: 6 } as JSONSchemaNumeric);
 
       expect(test).toBe(true);
 
@@ -578,7 +568,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with an integer equal to exclusiveMinimum`, () => {
 
-      const test = jsonValidator.validate(5, { type: 'integer', exclusiveMinimum: 5 });
+      const test = jsonValidator.validate(5, { type: 'integer', exclusiveMinimum: 5 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -586,7 +576,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with an integer less than an exclusiveMinimum`, () => {
 
-      const test = jsonValidator.validate(4, { type: 'integer', exclusiveMinimum: 5 });
+      const test = jsonValidator.validate(4, { type: 'integer', exclusiveMinimum: 5 } as JSONSchemaNumeric);
 
       expect(test).toBe(false);
 
@@ -598,7 +588,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false if data is not an object`, () => {
 
-      const test = jsonValidator.validate('', { required: [], properties: {} });
+      const test = jsonValidator.validate('', { required: [], properties: {} } as JSONSchemaObject);
 
       expect(test).toBe(false);
 
@@ -607,7 +597,7 @@ describe(`JSONValidator`, () => {
     it(`should throw if required properties are not described in 'properties'`, () => {
 
       expect(() => {
-        jsonValidator.validate({ test: 'test' }, { required: ['test'] });
+        jsonValidator.validate({ test: 'test' }, { properties: { other: { type: 'string' } }, required: ['test'] } as JSONSchemaObject);
       }).toThrowError();
 
     });
@@ -620,7 +610,7 @@ describe(`JSONValidator`, () => {
           test1: { type: 'string' },
           test2: { type: 'string' }
         }
-      });
+      } as JSONSchemaObject);
 
       expect(test).toBe(true);
 
@@ -635,7 +625,7 @@ describe(`JSONValidator`, () => {
           test2: { type: 'string' },
           test3: { type: 'string' }
         }
-      });
+      } as JSONSchemaObject);
 
       expect(test).toBe(true);
 
@@ -644,10 +634,11 @@ describe(`JSONValidator`, () => {
     it(`should return false on an object with missing required properties`, () => {
 
       const test = jsonValidator.validate({ test1: '' }, { required: ['test1', 'test2'],
-      properties: {
-        test1: { type: 'string' },
-        test2: { type: 'string' }
-      } });
+        properties: {
+          test1: { type: 'string' },
+          test2: { type: 'string' }
+        }
+      } as JSONSchemaObject);
 
       expect(test).toBe(false);
 
@@ -659,7 +650,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false if data is not an object`, () => {
 
-      const test = jsonValidator.validate('', { properties: {} });
+      const test = jsonValidator.validate('', { properties: {} } as JSONSchemaObject);
 
       expect(test).toBe(false);
 
@@ -676,7 +667,7 @@ describe(`JSONValidator`, () => {
             type: 'number'
           }
         }
-      });
+      } as JSONSchemaObject);
 
       expect(test).toBe(true);
 
@@ -693,7 +684,7 @@ describe(`JSONValidator`, () => {
             type: 'number'
           }
         }
-      });
+      } as JSONSchemaObject);
 
       expect(test).toBe(false);
 
@@ -710,7 +701,7 @@ describe(`JSONValidator`, () => {
             type: 'number'
           }
         }
-      });
+      } as JSONSchemaObject);
 
       expect(test).toBe(false);
 
@@ -732,7 +723,7 @@ describe(`JSONValidator`, () => {
             required: ['test3']
           }
         }
-      });
+      } as JSONSchemaObject);
 
       expect(test).toBe(true);
 
@@ -754,7 +745,7 @@ describe(`JSONValidator`, () => {
             required: ['test3']
           }
         }
-      });
+      } as JSONSchemaObject);
 
       expect(test).toBe(false);
 
@@ -762,7 +753,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with valid arrays nested in objects`, () => {
 
-      const test = jsonValidator.validate({ test: ['', ''] }, { properties: { test: { items: { type: 'string' } } } });
+      const test = jsonValidator.validate({ test: ['', ''] }, { properties: { test: { items: { type: 'string' } } } } as JSONSchemaObject);
 
       expect(test).toBe(true);
 
@@ -772,17 +763,9 @@ describe(`JSONValidator`, () => {
 
   describe(`validateItems`, () => {
 
-    it(`should throw if items does not contain type`, () => {
-
-      expect(() => {
-        jsonValidator.validate(['test'], { items: {} });
-      }).toThrowError();
-
-    });
-
     it(`should return false if data is not an array`, () => {
 
-      const test = jsonValidator.validate('', { items: { type: 'string' } });
+      const test = jsonValidator.validate('', { items: { type: 'string' } } as JSONSchemaArray);
 
       expect(test).toBe(false);
 
@@ -790,7 +773,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true if array items are valid`, () => {
 
-      const test = jsonValidator.validate(['', ''], { items: { type: 'string' } });
+      const test = jsonValidator.validate(['', ''], { items: { type: 'string' } } as JSONSchemaArray);
 
       expect(test).toBe(true);
 
@@ -798,7 +781,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false if array items are invalid`, () => {
 
-      const test = jsonValidator.validate(['', ''], { items: { type: 'number' } });
+      const test = jsonValidator.validate(['', ''], { items: { type: 'number' } } as JSONSchemaArray);
 
       expect(test).toBe(false);
 
@@ -806,7 +789,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with valid nested arrays`, () => {
 
-      const test = jsonValidator.validate([[''], ['']], { items: { items: { type: 'string' } } });
+      const test = jsonValidator.validate([[''], ['']], { items: { items: { type: 'string' } } } as JSONSchemaArray);
 
       expect(test).toBe(true);
 
@@ -814,7 +797,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with invalid nested arrays`, () => {
 
-      const test = jsonValidator.validate([[''], ['']], { items: { items: { type: 'number' } } });
+      const test = jsonValidator.validate([[''], ['']], { items: { items: { type: 'number' } } } as JSONSchemaArray);
 
       expect(test).toBe(false);
 
@@ -822,7 +805,9 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with valid objects nested in arrays`, () => {
 
-      const test = jsonValidator.validate([{ test: 'test' }, [{ test: 'test' }]], { items: { properties: { test: { type: 'string' } } } });
+      const test = jsonValidator.validate([{ test: 'test' }, [{ test: 'test' }]], {
+        items: { properties: { test: { type: 'string' } } }
+      } as JSONSchemaArray);
 
       expect(test).toBe(true);
 
@@ -831,7 +816,7 @@ describe(`JSONValidator`, () => {
     it(`should throw if maxItems is not an integer`, () => {
 
       expect(() => {
-        jsonValidator.validate([], { items: { type: 'string' }, maxItems: 10.5 });
+        jsonValidator.validate([], { items: { type: 'string' }, maxItems: 10.5 } as JSONSchemaArray);
       }).toThrowError();
 
     });
@@ -839,14 +824,14 @@ describe(`JSONValidator`, () => {
     it(`should throw if maxItems is not positive`, () => {
 
       expect(() => {
-        jsonValidator.validate([], { items: { type: 'string' }, maxItems: -1 });
+        jsonValidator.validate([], { items: { type: 'string' }, maxItems: -1 } as JSONSchemaArray);
       }).toThrowError();
 
     });
 
     it(`should return true with an array with less items than maxItems`, () => {
 
-      const test = jsonValidator.validate([1, 2], { items: { type: 'number' }, maxItems: 3 });
+      const test = jsonValidator.validate([1, 2], { items: { type: 'number' }, maxItems: 3 } as JSONSchemaArray);
 
       expect(test).toBe(true);
 
@@ -854,7 +839,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with an array length equal to maxItems`, () => {
 
-      const test = jsonValidator.validate([1, 2], { items: { type: 'number' }, maxItems: 2 });
+      const test = jsonValidator.validate([1, 2], { items: { type: 'number' }, maxItems: 2 } as JSONSchemaArray);
 
       expect(test).toBe(true);
 
@@ -862,7 +847,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with an array with more items than maxItems`, () => {
 
-      const test = jsonValidator.validate([1, 2, 4, 4], { items: { type: 'number' }, maxItems: 3 });
+      const test = jsonValidator.validate([1, 2, 4, 4], { items: { type: 'number' }, maxItems: 3 } as JSONSchemaArray);
 
       expect(test).toBe(false);
 
@@ -871,7 +856,7 @@ describe(`JSONValidator`, () => {
     it(`should throw if minItems is not an integer`, () => {
 
       expect(() => {
-        jsonValidator.validate([], { items: { type: 'string' }, minItems: 10.5 });
+        jsonValidator.validate([], { items: { type: 'string' }, minItems: 10.5 } as JSONSchemaArray);
       }).toThrowError();
 
     });
@@ -879,14 +864,14 @@ describe(`JSONValidator`, () => {
     it(`should throw if minItems is not positive`, () => {
 
       expect(() => {
-        jsonValidator.validate([], { items: { type: 'string' }, minItems: -1 });
+        jsonValidator.validate([], { items: { type: 'string' }, minItems: -1 } as JSONSchemaArray);
       }).toThrowError();
 
     });
 
     it(`should return true with an array with more items than minItems`, () => {
 
-      const test = jsonValidator.validate([1, 2, 3, 4], { items: { type: 'number' }, minItems: 3 });
+      const test = jsonValidator.validate([1, 2, 3, 4], { items: { type: 'number' }, minItems: 3 } as JSONSchemaArray);
 
       expect(test).toBe(true);
 
@@ -894,7 +879,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with an array length equal to minItems`, () => {
 
-      const test = jsonValidator.validate([1, 2], { items: { type: 'number' }, minItems: 2 });
+      const test = jsonValidator.validate([1, 2], { items: { type: 'number' }, minItems: 2 } as JSONSchemaArray);
 
       expect(test).toBe(true);
 
@@ -902,7 +887,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with an array with less items than maxItems`, () => {
 
-      const test = jsonValidator.validate([1, 2], { items: { type: 'number' }, minItems: 3 });
+      const test = jsonValidator.validate([1, 2], { items: { type: 'number' }, minItems: 3 } as JSONSchemaArray);
 
       expect(test).toBe(false);
 
@@ -910,7 +895,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true with an array with uniqueItems`, () => {
 
-      const test = jsonValidator.validate([1, 2], { items: { type: 'number' }, uniqueItems: true });
+      const test = jsonValidator.validate([1, 2], { items: { type: 'number' }, uniqueItems: true } as JSONSchemaArray);
 
       expect(test).toBe(true);
 
@@ -918,7 +903,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false with an array with non uniqueItems`, () => {
 
-      const test = jsonValidator.validate([1, 1], { items: { type: 'number' }, uniqueItems: true });
+      const test = jsonValidator.validate([1, 1], { items: { type: 'number' }, uniqueItems: true } as JSONSchemaArray);
 
       expect(test).toBe(false);
 
@@ -930,7 +915,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false if array length is not equel to schemas length`, () => {
 
-      const test = jsonValidator.validate(['', 10], { items: [{ type: 'string' }] });
+      const test = jsonValidator.validate(['', 10], { items: [{ type: 'string' }] } as JSONSchemaArray);
 
       expect(test).toBe(false);
 
@@ -938,7 +923,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return true if array values match schemas`, () => {
 
-      const test = jsonValidator.validate(['', 10], { items: [{ type: 'string' }, { type: 'number' }] });
+      const test = jsonValidator.validate(['', 10], { items: [{ type: 'string' }, { type: 'number' }] } as JSONSchemaArray);
 
       expect(test).toBe(true);
 
@@ -946,7 +931,7 @@ describe(`JSONValidator`, () => {
 
     it(`should return false if array values mismatch schemas`, () => {
 
-      const test = jsonValidator.validate(['', 10], { items: [{ type: 'string' }, { type: 'boolean' }] });
+      const test = jsonValidator.validate(['', 10], { items: [{ type: 'string' }, { type: 'boolean' }] } as JSONSchemaArray);
 
       expect(test).toBe(false);